requirements.txt
websockets>=12.0
python-binance>=1.0.19
